Refactor metric parsing
